requirements.txt
appdirs==1.4.0
Django==1.10.5
django-debug-toolbar==1.6
djangorestframework==3.5.3
djangorestframework-jwt==1.9.0
djoser==0.5.4
httpie==0.9.9
packaging==16.8
psycopg2==2.6.2
PyJWT==1.4.2
pyparsing==2.1.10
requests==2.13.0
six==1.10.0
sqlparse==0.2.2
